Externalize the looping over Sheets
Added methods that take a list of SheetInfo object, rather than working
with an internally derived list of Sheets.  These methods also return
two lists, the SheetInfo objects for the Sheets that were updated, and
the SheetInfo objects (along with a bit of additional info), for those
Sheets that were not updated.

This will enable retrying of bulk operations in the event of
Sheet-specific failures.